convert scheduler to use absolute ns instead of relative ticks
The scheduler now supports a ns granularity (vs the old 10MHz tick) in anticipation of having multiple clocked devices at different time bases.  As a result, times are now int64 types, and all the complicated wrapping logic is gone.

+++ src/IoCardDisplay.cpp
@@ -285,7 +285,7 @@ IoCardDisplay::tcbHsync(int arg)
 
     m_hsync_count++;
 
-    int new_period;
+    int64 new_period;
     if (regulated && (m_hsync_count < num_scanlines)) {
         // end of horizontal sweep; do another
         new_period = HSYNC_PERIOD;
